fixed a build error with authOptions, refactored authOptions so we can use the session on server
